// src/enums/actions.ts
export enum SoMediaManagerAction {
  // SELECT_END = "SELECT_END",
  CREATE_FOLDER = "CREATE_FOLDER",
  OPEN_FOLDER = "OPEN_FOLDER",
  OPEN_FILE = "OPEN_FILE",
  SEARCH = "SEARCH",
  UPLOAD_FILE = "UPLOAD_FILE",
  MOVE_SELECTION = "MOVE_SELECTION",
  DOWNLOAD_SELECTION = "DOWNLOAD_SELECTION",
  DELETE_SELECTION = "DELETE_SELECTION",
  LIST_VIEW = "LIST_VIEW",
  GRID_VIEW = "GRID_VIEW",
  SORT_ACTION = "SORT_ACTION",
  SORT_NAME = "SORT_NAME",
  SORT_SIZE = "SORT_SIZE",
  SORT_DATE = "SORT_DATE",
  SORT_DIRECTION_ASC = "SORT_DIRECTION_ASC",
  SORT_DIRECTION_DESC = "SORT_DIRECTION_DESC",
}